Made ComboBox a top level concept
